Added the images and R codes associated with Project 1 of the Coursera's Exploratory Data Analysis, course 4 of the JHU's Data Science Specialization.
Note that the working directory specified in the code (setwd("D://COURSE..)
is where the dataset is stored on my PC and not necessarily on yours too.
Before running the code, please ensure to change the working directory
depending on where the dataset is saved in your PC. This directory will
also be the destination of the png images that will be generated.
I know it's quite obvious upon seeing the code but just in case it's not..